Functional command line game of tic tac toe
This tic-tac-toe game is played in the command line. Built in order to
practice working with classes, the game consists of 3 main classes:
A Square class to represent each playable tile, a GameBoard class to
represent a collection of 9 Squares, and a GameRound class that controls
the game flow, accepts and validates user input, keeps track of turns and
selected tiles, and ulitmately checks for a winning condition of 3 X's or
3 O's in a row. All 8 win conditions(3 horizontal, 3 vertical, and 2
diagonal) are valid. Currently, the game exits upon a win or draw.