Initial commit. Functional one-way Mastermind game
This version of the game Mastermind allows the user be the guesser
playing against the computer. For the command line version of this game,
the computer generates a 4 digit code consisting of the numbers 1-6.
The user has 12 turns to try and guess the code. After each turn, the
computer provides feedback in the form of three possible symbols. A 'X'
represents a correct guess of both value and position of a digit in the
code. A '~' represents a digit value that is correct, but currently in
the wrong position. A '.' represents a value that in not found anywhere
in the code. Each round of feedbackis printed to the screen in an array.
This array is position indepent of the user guess, and simply represents
the number of exact, partial, and incorrect matches.

Future functionality could include allowing the user and having the
computer "player" attempt to solve the code. Improvements in code could
include factoring of code into additional classes and/or subclasses.