Fully functional CL Hangman w/ save game feature
This command line version of the classic Hangman game demonstrates use
of classes to handle game function, as well as serialization in order
to allow the user to save their game. All neccessary game instance
variables are saved to a yml file, to be loaded later, allowing the user
to continue a previously saved game.